# swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local server"
    }
  ],
  "info": {
    "version": "1.0.0",
    "title": "MTO Backend Project"
  },
...
  "paths": {
...
    "/admin/products/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "description": "Get the specific product with the given id",
        "operationId": "getProductById",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "id must be a string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No data found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataFoundMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/merchant/products": {
      "get": {
        "tags": [
          "merchant"
        ],
        "description": "Get all products which are belong to a specific merchant",
        "operationId": "getProductsByMerchant",
        "parameters": [
          {
            "$ref": "#/components/parameters/secretToken"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/fromPrice"
          },
          {
            "$ref": "#/components/parameters/toPrice"
          },
          {
            "$ref": "#/components/parameters/shopAddress"
          },
          {
            "$ref": "#/components/parameters/name"
          },
          {
            "$ref": "#/components/parameters/sortBy"
          },
          {
            "$ref": "#/components/parameters/order"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidFilterMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMessage"
                }
              }
            }
          },
          "404": {
            "description": "No data found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataFoundMessage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "merchant"
        ],
        "description": "Create a new product",
        "operationId": "createProduct",
        "parameters": [
          {
            "$ref": "#/components/parameters/secretToken"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/NewProduct"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "name must be a string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMessage"
                }
              }
            }
          }
        }
      }
    },
    "/merchant/products/{id}": {
      "get": {
        "tags": [
          "merchant"
        ],
        "description": "Get the specific product with the given id",
        "operationId": "getProductByMerchantById",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/secretToken"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "id must be a number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMessage"
                }
              }
            }
          },
          "404": {
            "description": "No data found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataFoundMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "merchant"
        ],
        "description": "update the specific product with the given id",
        "operationId": "updateProductByMerchantById",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/secretToken"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/NewProduct"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "name must be a string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMessage"
                }
              }
            }
          },
          "404": {
            "description": "No data found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Cannot find the data to update"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "merchant"
        ],
        "description": "delete the specific product with the given id",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/secretToken"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Product removed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "id must be a string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMessage"
                }
              }
            }
          },
          "404": {
            "description": "No data found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Cannot find the data to delete"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "id": {
        "in": "path",
        "name": "id",
        "schema": {
          "type": "integer"
        },
        "required": true
      },
      "secretToken": {
        "in": "header",
        "name": "secret_token",
        "schema": {
          "type": "string"
        },
        "required": true
      },
...
      "page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "number"
        }
      },
      "limit": {
        "in": "query",
        "name": "limit",
        "schema": {
          "type": "number"
        }
      },
      "fromPrice": {
        "in": "query",
        "name": "fromPrice",
        "schema": {
          "type": "number"
        }
      },
      "toPrice": {
        "in": "query",
        "name": "toPrice",
        "schema": {
          "type": "number"
        }
      },
...
      "shopAddress": {
        "in": "query",
        "name": "shopAddress",
        "schema": {
          "type": "string"
        }
      },
      "name": {
        "in": "query",
        "name": "name",
        "schema": {
          "type": "string"
        }
      },
...
      "sortBy": {
        "in": "query",
        "name": "sortBy",
        "schema": {
          "type": "string"
        }
      },
      "order": {
        "in": "query",
        "name": "order",
        "schema": {
          "type": "string",
          "enum": [
            "DESC",
            "ASC"
          ]
        }
      }
    },
    "schemas": {
...
      "Merchant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "walletAddress": {
            "type": "string",
            "example": "0x12345abcde7890"
          },
          "shippingAddress": {
            "type": "string"
          },
          "externalLink": {
            "type": "string"
          }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "merchant": {
            "type": "integer",
            "allOf": [
              {
                "$ref": "#/components/schemas/Merchant"
              }
            ]
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "shopAddress": {
            "type": "string"
          }
        }
      },
      "UnauthorizedMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid JWT"
          }
        }
      },
      "NoDataFoundMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "No data found"
          }
        }
      },
      "InvalidFilterMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "page must be greater than or equal to 1"
          }
        }
      }
    },
    "requestBodies": {
...
      "NewProduct": {
        "description": "Product object that needs to be added to the store",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "shopAddress": {
                  "type": "string"
                }
              }
            }
          }
        },
        "required": true
      }
    }
  }
}